# Provider sends 1st invoice wo delay