--- a/core/service/session_manager_test.go
+++ b/core/service/session_manager_test.go
@@ -35,6 +35,7 @@ import (
 	"github.com/mysteriumnetwork/node/pb"
 	sessionEvent "github.com/mysteriumnetwork/node/session/event"
 	"github.com/mysteriumnetwork/node/trace"
+	"github.com/mysteriumnetwork/payments/crypto"
 	"github.com/stretchr/testify/assert"
 )
 
@@ -317,7 +318,7 @@ func newManager(service *Instance, sessions *SessionPool, publisher publisher, p
 	return NewSessionManager(
 		service,
 		sessions,
-		func(_, _ identity.Identity, _ common.Address, _ string) (PaymentEngine, error) {
+		func(_, _ identity.Identity, _ common.Address, _ string, _ chan crypto.ExchangeMessage) (PaymentEngine, error) {
 			return paymentEngine, nil
 		},
 		&MockNatEventTracker{},