--- a/session/pingpong/exchange_messaging.go
+++ b/session/pingpong/exchange_messaging.go
@@ -19,7 +19,6 @@ package pingpong
 
 import (
 	"context"
-	"fmt"
 	"time"
 
 	"github.com/mysteriumnetwork/node/p2p"
@@ -68,36 +67,4 @@ func (es *ExchangeSender) Send(em crypto.ExchangeMessage) error {
 	defer cancel()
 	_, err := es.ch.Send(ctx, p2p.TopicPaymentMessage, p2p.ProtoMessage(pMessage))
 	return err
-}
-
-func exchangeMessageReceiver(channel p2p.ChannelHandler) (chan crypto.ExchangeMessage, error) {
-	exchangeChan := make(chan crypto.ExchangeMessage, 1)
-
-	channel.Handle(p2p.TopicPaymentMessage, func(c p2p.Context) error {
-		var msg pb.ExchangeMessage
-		if err := c.Request().UnmarshalProto(&msg); err != nil {
-			return fmt.Errorf("could not unmarshal exchange message proto: %w", err)
-		}
-		log.Debug().Msgf("Received P2P message for %q: %s", p2p.TopicPaymentMessage, msg.String())
-
-		exchangeChan <- crypto.ExchangeMessage{
-			Promise: crypto.Promise{
-				ChannelID: msg.GetPromise().GetChannelID(),
-				Amount:    msg.GetPromise().GetAmount(),
-				Fee:       msg.GetPromise().GetFee(),
-				Hashlock:  msg.GetPromise().GetHashlock(),
-				R:         msg.GetPromise().GetR(),
-				Signature: msg.GetPromise().GetSignature(),
-			},
-			AgreementID:    msg.GetAgreementID(),
-			AgreementTotal: msg.GetAgreementTotal(),
-			Provider:       msg.GetProvider(),
-			Signature:      msg.GetSignature(),
-			HermesID:       msg.GetHermesID(),
-		}
-
-		return nil
-	})
-
-	return exchangeChan, nil
 }