Optimising provider lag on session establishment by removing sleep before provider sends 1st invoice. Also subscribe P2P payment messages earlier.

Before:
```
Provider connection trace: "Provider whole session create" took 1.47595767s, "Provider session start" took 31.068551ms, "Provider payments" took 1.328130467s, "Provider config" took 116.749528ms
```

After:
```
Provider connection trace: "Provider whole session create" took 386.80665ms, "Provider session start" took 33.366328ms, "Provider payments" took 273.28457ms, "Provider config" took 80.151728ms
```