@vkuznecovas 2020-08-20
Have you tested this with a connection that's a bit slower? I'm pretty sure that will fail.